+++ changelog.md
@@ -1,4 +1,13 @@
-#GPS track record Visualizer Changelog
+# GPS track record Visualizer Changelog
+
+## Version 1.1.1
+1. UI improvement
+2. Highlight plot on Map when clicked on chart and vice-versa
+
+## Version 1.1.0
+1. Track Info section added
+2. Head NavBar added
+3. UI improvement
 
 ## Version 1.0.2
 1. Demo mode enabled